refactor: move reusable code to separate space
