+++ app/apps/base/models.py
@@ -1,0 +1,13 @@
+from django.db import models
+import uuid
+
+# Models are defined here
+
+class TimeStampedUUIDModel(models.Model):
+    id = models.UUIDField(primary_key=True, editable=False, default=uuid.uuid4)
+    created_at = models.DateTimeField(auto_now_add=True)
+    modified_at = models.DateTimeField(auto_now=True)
+
+    class Meta:
+        abstract = True
+
